{{cookiecutter.repo_name}}/requirements/common.txt
# Bleeding edge Django
django==1.9

# Configuration
django-environ==0.4.0
django-secure==1.0.1

{% if cookiecutter.use_white_red_noise == 'white' %}
whitenoise==2.0.6
{% elif cookiecutter.use_white_red_noise == 'red' %}
whitenoise==2.0.6
django-rednoise==1.0.5
{%- endif %}

